default to Regular font with FontConfig
Default to Regular/Normal instead of Medium in QFontConfigDatabase.

Requesting Medium as default was fine until the font actually provided a
Medium font, at which point it was selected by default.

Tweak the buckets so that Medium has preference over Bold when DemiBold
is requested. They will have to be extended for Medium when QTBUG-25814
is resolved.

Change-Id: I111b7f99083d50c3e8264f84f10a997f619ef4b3
Reviewed-by: Konstantin Ritt <ritt.ks@gmail.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>