QNX: Fix crash due to re-entering QToolTip::show().
The flushWindowSystemEvents() was added to send the geometry
event synchronously. A side effect was that all input events were
also sent.

We don't want that. We're probably in setGeometry() because some
other input event brought us here, and it's still in the call
stack.

One crashy example is the static method QToolTip::show() which
uses static variables([1]). If we process input events while the
first QToolTip::show() is still on the call stack, it will trigger
a second one, deleting stuff ([1]) that the first invocation will
still need once the stack unwinds.

Change-Id: Iea22419a7f386d6b6231404666092340dfd1afed
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>