Remove Plastique and Cleanlooks from the widget gallery
This commit does not remove the actual images, since they
are referenced from many other places.

Plastique style seems to have been used often as a style
for showing a screenshot of a widget (probably because it
has been available on all platforms).

When fusion style gets its gallery with images, the references
to the plastique images should probably be replaced with
references to fusion images.

Change-Id: I288c06df2c5e1f3a212b873ec0f087e119b895b0
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>