Examples: inherit NorwegianWoodStyle from QProxyStyle
The QStyle specializations are being made internal. The recommended way
to customize styles is now to use QProxyStyle (& QStyleFactory), or to
implement a full custom style one can alternatively subclass
QCommonStyle. The proxy style approach was chosen for this case, since
the example assumes some drawing functionality provided by the windows
style and it is not a "complete" custom style implementation but more
like a customization anyway.

Change-Id: Ib2477339cfef258cfc944a76a2eea728066e1f45
Reviewed-by: Geir Vattekar <geir.vattekar@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>