QtBase: Fix painting example arthur background color
The arthur style set the background color to a bright color
and leave the foregroud color, for example for text, untouched.

If you are using a color theme with bright foregroundcolors on your
system this will result in unreadable text on push- and
radio-buttons.

The function drawControl is now reimplemented and the labels
with text are drawn with an appropriate color.

Change-Id: Ifd7f3a2b1c0305130255009c7733e0966830171a
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>